+++ packages/atomic/src/components/facets-v1/atomic-timeframe/timeframe.ts
@@ -1,0 +1,5 @@
+import {RelativeDate} from '@coveo/headless';
+
+export interface Timeframe extends RelativeDate {
+  label?: string;
+}