+++ packages/samples/headless-react/src/components/date-facet/date-facet.class.tsx
@@ -39,7 +39,7 @@ export class DateFacet extends Component<DateFacetProps, DateFacetState> {
   }
 
   private getKeyForRange(value: DateFacetValue) {
-    return `[${value.start}..${value.end}${value.endInclusive ? ']' : '['}`;
+    return `[${value.start}..${value.end}]`;
   }
 
   private format(dateStr: string) {
@@ -51,7 +51,11 @@ export class DateFacet extends Component<DateFacetProps, DateFacetState> {
       return null;
     }
 
-    if (!this.state.values.length) {
+    if (
+      !this.state.values.filter(
+        (value) => value.state !== 'idle' || value.numberOfResults > 0
+      ).length
+    ) {
       return <div>No facet values</div>;
     }
 
@@ -65,8 +69,7 @@ export class DateFacet extends Component<DateFacetProps, DateFacetState> {
               onChange={() => this.controller.toggleSelect(value)}
               disabled={this.state.isLoading}
             />
-            {this.format(value.start)} to {this.format(value.end)}{' '}
-            {value.endInclusive ? 'inclusively' : 'exclusively'} (
+            {this.format(value.start)} to {this.format(value.end)} (
             {value.numberOfResults}{' '}
             {value.numberOfResults === 1 ? 'result' : 'results'})
           </li>