+++ packages/headless/CHANGELOG.md
@@ -3,6 +3,67 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.11.3](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.11.2...@coveo/headless@1.11.3) (2021-07-29)
+
+
+### Bug Fixes
+
+* **headless:** set hasResults to false when hasError is true ([#1014](https://github.com/coveo/ui-kit/issues/1014)) ([d4a8254](https://github.com/coveo/ui-kit/commit/d4a8254f16e1aa6d2f84915e09bba9f4db622059))
+
+
+
+
+
+## [1.11.2](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.11.1...@coveo/headless@1.11.2) (2021-07-28)
+
+**Note:** Version bump only for package @coveo/headless
+
+
+
+
+
+## [1.11.1](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.11.0...@coveo/headless@1.11.1) (2021-07-28)
+
+**Note:** Version bump only for package @coveo/headless
+
+
+
+
+
+# [1.11.0](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.10.0...@coveo/headless@1.11.0) (2021-07-28)
+
+
+### Features
+
+* **headless:** add timezone to engines config, deprecate useLocalTime for buildDateRange ([#1006](https://github.com/coveo/ui-kit/issues/1006)) ([abe64be](https://github.com/coveo/ui-kit/commit/abe64be498600e0d722aef6d1a2c34a8fc171f98))
+
+
+
+
+
+# [1.10.0](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.9.0...@coveo/headless@1.10.0) (2021-07-28)
+
+
+### Features
+
+* **atomic:** support Facet Value Captions ([#997](https://github.com/coveo/ui-kit/issues/997)) ([aa9bc6e](https://github.com/coveo/ui-kit/commit/aa9bc6eb71b71367c934b14f11e80cdc5777a754))
+* **headless:** expose synchronize method on search parameter manager ([#1000](https://github.com/coveo/ui-kit/issues/1000)) ([a100a7f](https://github.com/coveo/ui-kit/commit/a100a7f32aad8af07c2b18d40a6a7efbe5fc9ec0))
+
+
+
+
+
+# [1.9.0](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.8.0...@coveo/headless@1.9.0) (2021-07-26)
+
+
+### Features
+
+* **headless:** support relative dates in date facet ([#987](https://github.com/coveo/ui-kit/issues/987)) ([b4d8ad8](https://github.com/coveo/ui-kit/commit/b4d8ad8ffdb554930f15c350f25d4a49d77915d6))
+
+
+
+
+
 # [1.8.0](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.7.4...@coveo/headless@1.8.0) (2021-07-23)